Add example cert config file
The certificat config file, with a format established in another set
of scripts is expected to have an all uppercase name, with lowercase
.cfg extension. The file is to be sourced in Bash and the created
variable names have the same prefix, all caps, as the file name.

The other scripts update the .cfg file with the full fingerprints of
two generated keys, one RSA and one ECC, which share the remaining
data as if they were one certificate.

This file is sourced by this script to determine file names, comment
field contents and to select which certificate to process.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>