Change name of project
The initial script for the project is for splitting certficates into
files for export.

There will be other purposes, designed around the managment of GnuPG
certificates. The name change is to recognize that this will be a
collection of tools for GnuPG management, rather than just the first
such tool.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>