Switch to passed parameters in functions
Even though BASH uses generally global variables, switch to using the
parameter list to pass data into the functions.

Replaced the load_config_data with check_config_data. Using the
function to obtain a "return" value placed the global data in an
unreachable sub-shell and it was lost. All the function does now is
verify that the config file exists, and errors if it doesn't.

The data which was loaded will be loaded closer to where it is used.

The comments after the function names will be dropped in the final
version and are there now to remind me what is being passed in by
the caller.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>