Add routine to handle extracting the primary key
Extract the primary key to a variety of files and formats. Using the
current (management) passphrase to be kept on the certificate system,
using a backup passphrase for cold-storage, in both ASCII armored and
GPG native formats. Also a public key for export, if/when it's needed
for something.

At this point the writing is stubbed, while system logic is developed
and will be implemented later.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>