Split notices from real errors
Modify the error notice routine to indicate an error has occured, and
add "Aborting" automatically. Add a new routine to use when there's
nothing to do. Still an abort from the program stand-point, but not a
system error. Modify the messages accordingly.

Split the longer commands into continuation lines. Functions called as
sub-shells `$"(...)` are also split and indented. More for continuity
than for any programmatic reasons.

Removed some of the trace outputs for sections which I believe are in
a final, or near final state.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>