Reduce namespace pollution
Bash uses variable names as global by default. Using names inside the
functions when not needed is causing side effects and polluting the
available name space. Switched to using the postitional parameters
when ever possible and when not possible used disposable names which
are used in their routine, and ignored in other locations.

Since the functions should only be called from within the script, I
don't need to keep checking for missing parameters on the call, and
the in routine testing of them is also dropped.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>