Unify the file writing processes
The routines for extracting the keys and certs to files were similar
enough to prompt me to reduce the process to a generic routine and
have the parameteres passed to control what was exported and how the
file was created.

Updated the .config file to control the directories to use for saving
the created files.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>