Change screen display of progress
Removed the tracing output from the finished routines. Removed the
parameter list notes from finished routines. I found that the screen
was alarmingly silent as progress was made.

Added an output to the user for each cert or key saved to a file.
This seems to keep the screen from staying silent for too long.

Making the change for a usable screen output included resetting the
sequence of parameters to the routines so that they are in the same
order, with additions to the end, across the levels used. A few of
the variable names were changed to help with this and newlines were
prefixed to error messages, as they will now be displayed where a
progress message is half written.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>