Add routine to search through a passphrase list
Now that there is the ability to test a passphrase for validity, we
need to run through a list of possible ones to see which, if any, of
them is the one needed for the certificate. Saving the found one will
also be needed. If none of them validates, then we need to error out
early.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>