Include the reserved keys in the exports
Each certificate includes a key in the alternate algorithm for each
of the purposes. An ECC key for GIT use has an auth and sign key in
ECC, with a matching auth and sign key in RSA. These 'reserve' keys
are not included in the extraction process. Adding them included a
need to expand the file name format. The status messages are also
modified to indicate which one is being saved, or has failed.

The last of the diagnostic and programming aids have been removed as
well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>