Add example files
Replacing the original sample config file with a new one which is
part of a set of sample files, and which was used to generate a
complete set of certificates, contained in `SAMPLE.keys` so as to
have a accurate example, with real data, to show.

The files `SAMPLE*.cfg` and are examples of the config files which
certspliter will look for. They are generated from a different script
and the sequence of the entries does not matter.

The file name, not counting the `.cfg` extension **must** be all caps
and the variable names inside **must** match, in case and spelling.
The `*_pfx`, `*_ecc_sec`, `*_rsa_sec`, `*_user` and `*_primary_uid`
variables are the only ones used by certspliter.

The other script uses the same files to generate a certificate using
both RSA and ECC algorithms. Each key has the primary key set for
[C] certification use only with the first subkey, using a matching
algorithm, set for the use given in the config file. Other scripts
will add subkeys as needed, with each subkey being part of a pair of
keys, one matching the algorithm of the primary key, and a partner
key using the opposite (RSA vs ECC) algorithm.

While it is possible for the other scripts to create more than one
group of keys for any given purpose, and to add keys which are [SA]
signing and authentication, certsplit will only process the last such
key for a given purpose and will not process [SA] subkeys at all.

The certsplitter script can be used to process any GPG certificate
made in any manner. However, if the processed key does not follow the
same pattern as used in the examples in `SAMPLE.keys` the results are
completely undefined. Error detection built into the script _should_
catch such behavior and provide intelligent, though perhaps not so
helpful, error messages.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>