Add sample output tree
Using the keys, shown in `SAMPLE.keys`, generated by the given
`SAMPLE*.cfg` files, provides the output files, in the tree as
given in the `SAMPLE_files.md`

Of note is that the directories `extracts` and `imports` will share a
pair of filenames, `*.ecc.cert.sec.gpg` and `*.rsa.cert.sec.gpg`.
Both files have the same complete certificate, with all secret keys,
but they have different passphrases. The `extracts` directory uses
the same passphrase as the original in the keyring, while the files
in the `imports` directory have a modified passphrase, which is not
the same as the modified passphrase used for the subkeys in the same
directory.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>