Build keyimporter process
The keyimporter process is to take a keyring name and a cert config
prefix followed by a space separated list of purpose names and add
the secret key for each of the named purposes into the named keyring.
When no purposes are listed, the process will attempt to import each
of the three purposes (auth, sign, and crypt), silently ignoring any
which do not exist. If the purposes are listed, only purposes which
exist and match the named purposes are imported, and any missing ones
will cause a fatal error. In addition, if the cert config prefix is
for a GIT key or an SSH key, and the auth purpose is included, the
proper SSH-related files will be copied to where they are needed. If
the auth key is not imported, the SSH-related files are not copied
by default. The copying of the SSH-related files can be forced, even
for non SSH or GIT keys, by using the purpose of 'ssh' in the list of
purposes. If SSH is forced, or if auth is named and the SSH-related
files are missing, a fatal error occurs. If the auth key is copied as
the default, rather than explicitly listed, and the SSH-related keys
are expected, but not found, a fatal error occurs.

Only keys under a primary key of the ECC variety are examened, and
only ECC subkeys are imported.

By default the primary key is NOT imported, and it cannot be imported
while the rest of the keys are being imported. To import the primary
key the command must be invoked specifying the single purpose of CERT
in all caps. In addition, all other settings to the contrary, the
passphrase for the primary key will be required from the terminal.
The "terminal" in this case can be either the CLI, or a GUI, based on
the normal settings for the active user.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>