Standardize the GPG options
Place the GPG options into some defined variables which can be used as
needed in any situation. Elimiates the need to always repeat them in
the code and allows for them to be changed universally rather than the
need to search out and update every instance of use.

Modified the email address in the license block to reflect the repo in
use rather than the Advent of Code address the block was copied with.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>