Create an install script for the tools
The vision for this script collection indicates that it will be big
enough to need the config settings saved locally somewhere. The trend
at the moment is the use the XDG Base Directory Specification:
   https://specifications.freedesktop.org/basedir-spec/latest/
for user-specific file locations. In addition, as the scripts are
mostly intended for user access, and require no system installation,
the extra (helper) script files will be placed in the user-specific
data directory tree.

There might be a more standard location for 'helper' scripts, and if
I ever find out I might convert to that. As my research thus far has
found, either it is expected that all user scripts are in ~/bin or
the "package" has its own directory in /usr/local and the needed
links are placed in /usr/local/bin. Nothing about a user script which
has helper scripts somewhere else. I could place all of them in the
user bin directory, but that seems to be a cluttering of that space,
and misleading in that the helper scripts are not intended to be
launched independently, as an entry in the ~/bin directory might
suggest.

As I intend to use other scripts, or sets of scripts, in a similar
fashion, and would like to have them all in a collected space, I'll
use the vendor sub-tree concept (which I encountered once and cannot
find now): $XDG_CONFIG_HOME/chindraba/gpg-tools/ for config files and
$XDG_DATA_HOME/chindraba/gpg-tools/ for data files. That way, for
another project, say fixvid, I can use the config directory as
$XDG_CONFIG_HOME/chindraba/fixvid/ and when the time comes to do a
backup, or to reinstall the system the $HOME/.config/chindraba/
directory will have everything I need in it. Anyone who uses my work
for their own purposes, will recieve the same benefit.

Further changes will be needed, including updates to the scripts for
the new file locations, renaming a few of the files, and updates to
documentation. All such changes will be in the next few commits.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>